Moved POJOs to Model Package.
